local aparece primeiro na lista de propriedades, modificado parser para dividir a latitude e longitude recebida do prolog por 100, adapatdos valores para funcionar melhor com valores de longitude e latitude real, removido ourm da lista de cidades porque ficava muito proximo de fatima, adicionada propriedade local a alguns dos sitios, parser mete sempre a primeira letra das strings em maiuscula

+++ netbeans/src/Mapa/Arco.java
@@ -13,17 +13,17 @@ import java.awt.Graphics;
  * @author Chalkos
  */
 public class Arco extends Figura{
-    protected int x1 = 0;
-    protected int x2 = 0;
-    protected int y1 = 0;
-    protected int y2 = 0;
+    protected double x1 = 0;
+    protected double x2 = 0;
+    protected double y1 = 0;
+    protected double y2 = 0;
     
     public Arco(){
         super();
         this.color = Mapa.normalLinhas;
     }
     
-    public Arco(int x1, int y1, int x2, int y2) {
+    public Arco(double x1, double y1, double x2, double y2) {
         this();
         this.x1 = x1;
         this.x2 = x2;
